gtkdatabox_regions: Allocate polygon data points on the stack
We only ever pass 4 points at once to gdk_draw_polygon. There is no need to
create the array which holds these points on the heap. By allocating it on the
stack and removing it from the private data for the class we can cleanup the
code quite a bit.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>